all compile, all works

+++ quotes-api/src/test/java/com/exchangeinformant/services/StockDbServiceImplTest.java
@@ -7,20 +7,17 @@ import com.exchangeinformant.model.Stock;
 import com.exchangeinformant.repository.InfoRepository;
 import com.exchangeinformant.repository.StockRepository;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.Month;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.when;
@@ -51,7 +48,7 @@ class StockDbServiceImplTest {
     void shouldGetStockByCode() {
         Stock stock = new Stock("AAPL", "Apple" ,"USD", new ArrayList<>(){
             {
-                add(new Info(1,34d,LocalDateTime.now(),"AAPL"));
+                add(new Info(1,BigDecimal.valueOf(34d),LocalDateTime.now(),"AAPL"));
             }
         });
 
@@ -77,7 +74,7 @@ class StockDbServiceImplTest {
         LocalDateTime actualTime = LocalDateTime.of(2021, Month.APRIL, 24, 14, 33, 48);
         Stock stock = new Stock("AAPL", "Apple" ,"USD", new ArrayList<>(){
             {
-                add(new Info(1,34d,actualTime,"AAPL"));
+                add(new Info(1, BigDecimal.valueOf(34d),actualTime,"AAPL"));
             }
         });
 
@@ -95,7 +92,7 @@ class StockDbServiceImplTest {
         codes.add("AAPL");
         Stock stock = new Stock("AAPL", "Apple" ,"USD", new ArrayList<>(){
             {
-                add(new Info(1,34d,LocalDateTime.now(),"AAPL"));
+                add(new Info(1,BigDecimal.valueOf(34d),LocalDateTime.now(),"AAPL"));
             }
         });
         List<Stock> stocks = new ArrayList<>();
